# hysav-api/openapi.yaml
openapi: 3.1.0
info:
  title: HySav API
  version: "1.0.0"
  description: >
    HySav | AI Monit backend. AI-subscription tracking, usage monitoring, and
    waste detection for small teams. All endpoints are under /api/v1. Auth is
    an opaque bearer token from /auth/register or /auth/login. Money is integer
    cents; timestamps are ISO-8601 strings.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /workspaces/{id}:
    get:
      tags: [workspaces]
      summary: Workspace details + members
      parameters: [{ $ref: "#/components/parameters/workspaceId" }]
      responses:
        "200": { description: Workspace with members }
        "404": { description: Not found or not a member }
    patch:
      tags: [workspaces]
      summary: Rename workspace (admin)
      parameters: [{ $ref: "#/components/parameters/workspaceId" }]
      responses:
        "200": { description: Updated }
  /workspaces/{id}/members/{userId}:
    patch:
      tags: [workspaces]
      summary: Change a member's role/title/color (admin)
      description: Demoting the workspace's only admin is refused.
      parameters:
        - { $ref: "#/components/parameters/workspaceId" }
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [admin, member] }
                title: { type: string }
                color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
      responses:
        "200": { description: Updated }
        "400": { description: Would remove the last admin }
    delete:
      tags: [workspaces]
      summary: Remove a member from the workspace (admin)
      description: >
        Detaches them from tools and prefs; their user account survives.
        Removing the only admin is refused.
      parameters:
        - { $ref: "#/components/parameters/workspaceId" }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
        "400": { description: Would remove the last admin }
  /workspaces/{id}/plan:
    get:
      tags: [billing]
      summary: Plan gate status (any member)
      description: >
        active/reason (paid | trial | expired) + trialEndsAt/paidUntil. Write
        endpoints (tool create/update, usage, imports, invites, integrations)
        return 402 with a clear message when expired; reads and spend-reducing
        actions (cancelling/deleting tools) always work.
      parameters: [{ $ref: "#/components/parameters/workspaceId" }]
      responses:
        "200": { description: Plan status }
...
  /workspaces/{id}/invites:
    post:
      tags: [workspaces]
      summary: Invite a teammate by email (admin)
      description: Sends an invite email and returns the invite link (7-day expiry).
      parameters: [{ $ref: "#/components/parameters/workspaceId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [admin, member], default: member }
      responses:
        "201": { description: Invite created; includes inviteLink }
...
  /workspaces/{id}/notification-prefs:
    get:
      tags: [notifications]
      summary: Current user's prefs for this workspace
      parameters: [{ $ref: "#/components/parameters/workspaceId" }]
      responses:
        "200": { description: wasteAlerts / renewalAlerts / weeklyDigest booleans }
    patch:
      tags: [notifications]
      summary: Update prefs (per user, per workspace)
      parameters: [{ $ref: "#/components/parameters/workspaceId" }]
      responses:
        "200": { description: Updated prefs }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    workspaceId:
      name: id
      in: path
      required: true
      schema: { type: string }
...